+++ src/variables/MigrationManagerVariable.php
@@ -1,6 +1,7 @@
 <?php
 
-namespace Craft;
+namespace firstborn\migrationmanager\variables;
+use firstborn\migrationmanager\helpers\MigrationManagerHelper;
 
 /**
  * Deploy Variable provides access to database objects from templates
@@ -10,48 +11,55 @@ class MigrationManagerVariable
     /**
      * @return string
      */
-    public function getName()
+    /*public function getName()
     {
         return Craft::$app->plugins->getPlugin('migrationManager')->getName();
-    }
+    }*/
 
     /**
      * @return array
      */
-    public function getNewMigrations()
+    /*public function getNewMigrations()
     {
         return Craft::$app->migrationManager_migrations->getNewMigrations();
-    }
+    }*/
 
     /**
      * @return array
      */
-    public function getAppliedMigrations()
+    /*public function getAppliedMigrations()
     {
         return Craft::$app->migrationManager_migrations->getAppliedMigrations();
-    }
+    }*/
 
     /**
      * @return array
      */
-    public function getAssetSources()
+    /*public function getAssetSources()
     {
         return Craft::$app->assetSources->getAllSources();
-    }
+    }*/
 
     /**
      * @return array
      */
-    public function getAssetTransforms()
+    /*public function getAssetTransforms()
     {
         return Craft::$app->assetTransforms->getAllTransforms();
-    }
+    }*/
 
     /**
      * @return array
      */
-    public function getAllTagGroups()
+    /*public function getAllTagGroups()
     {
         return Craft::$app->tags->getAllTagGroups();
+    }*/
+
+    /**
+     * @return boolean
+     */
+    public function isVersion($version){
+        return MigrationManagerHelper::isVersion($version);
     }
 }